requirements.txt
lightning
torchvision
torchmetrics
pathlib
matplotlib
albumentations
tqdm
wandb